Refactor submit logic. Submit on enter, newline on shift+enter
