--- src/views/chat/SendMessage/index.tsx
@@ -1,1 +1,0 @@
-export * from './sendMessage'
