+++ src/views/chat/prompt-input-box/index.tsx
@@ -1,0 +1,1 @@
+export * from './prompt-input-box'
