+++ src/views/chat/prompt-input-box/prompt-input-box.module.scss
@@ -10,3 +10,8 @@
 		flex-shrink: 0;
 	}
 }
+.textarea {
+	height: 2.5rem;
+	resize: none;
+	width: 100%;
+}
